pyproject.toml
[project]
name = "ai-grc-compliance"
version = "0.1.0"
requires-python = "==3.12.*"
dependencies = [
    "fastapi>=0.115.0",
    "uvicorn[standard]>=0.30.0",
    "pydantic>=2.10.0",
    "pydantic-settings>=2.7.0",
    "sqlalchemy>=2.0.35",
    "psycopg2-binary>=2.9.0",
    "alembic>=1.13.0",
    "anthropic>=0.40.0",
    "openpyxl>=3.1.5",
    "python-dotenv>=1.0.1",
    "python-multipart>=0.0.12",
    "slowapi>=0.1.9",
    "pytest>=8.3.0",
    "pytest-asyncio>=0.24.0",
    "httpx>=0.27.0",
    "PyJWT>=2.8.0"
]
